+++ Adj/Program.hs
@@ -1,1 +1,4 @@
-module Adj.Program where
+module Adj.Program (module Exports) where
+
+import Adj.Program.Instruction as Exports
+import Adj.Program.Construction as Exports