+++ Adj/Program/Construction.hs
@@ -1,0 +1,5 @@
+module Adj.Program.Construction where
+
+import Adj.Algebra.Product ((:*:))
+
+newtype Construction t a = Construction (a :*: t (Construction t a))